+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Changed the core contracts to emit events while communicating with Zeus instead of using return types.
     - why: events are more highly structured, can be parsed via the ZeusScript ABI, and don't require manual tuple parsing.
 - `zeus env show <env>` now prints all of the contracts and environment parameters associated with an environment.
+- Fixed a bug with `zeus deploy cancel` which require duplicate prompts.
+
 
 0.0.8:
 - `zeus test` now supports running a glob of tests (`zeus test --env testnet my/path/*.sol`)